+++ demo/lib/data.dart
@@ -1,5 +1,10 @@
 class CustomData {
-  const CustomData({this.name = '[your name]'});
+  const CustomData({
+    this.name = '[your name]',
+    this.testing = false,
+  });
 
   final String name;
+
+  final bool testing;
 }