+++ pdf/lib/src/widgets/text.dart
@@ -19,7 +19,9 @@ import 'dart:math' as math;
 import 'package:meta/meta.dart';
 
 import '../../pdf.dart';
+import '../pdf/font/arabic.dart' as arabic;
 import '../pdf/font/bidi_utils.dart' as bidi;
+import '../pdf/options.dart';
 import 'annotations.dart';
 import 'basic.dart';
 import 'document.dart';
@@ -658,6 +660,8 @@ class RichTextContext extends WidgetContext {
       '$runtimeType Offset: $startOffset -> $endOffset  Span: $spanStart -> $spanEnd';
 }
 
+typedef Hyphenation = List<String> Function(String word);
+
 class RichText extends Widget with SpanningWidget {
   RichText({
     required this.text,
@@ -668,6 +672,7 @@ class RichText extends Widget with SpanningWidget {
     this.textScaleFactor = 1.0,
     this.maxLines,
     this.overflow = TextOverflow.visible,
+    this.hyphenation,
   });
 
   static bool debug = false;
@@ -699,6 +704,8 @@ class RichText extends Widget with SpanningWidget {
   var _mustClip = false;
 
   List<InlineSpan>? _preprocessed;
+
+  final Hyphenation? hyphenation;
 
   void _appendDecoration(bool append, _TextDecoration td) {
     if (append && _decorations.isNotEmpty) {
@@ -923,9 +930,11 @@ class RichText extends Widget with SpanningWidget {
           final space =
               font.stringMetrics(' ') * (style.fontSize! * textScaleFactor);
 
-          final spanLines = (_textDirection == TextDirection.rtl
-                  ? bidi.logicalToVisual(span.text!)
-                  : span.text)!
+          final spanLines = (useArabic && _textDirection == TextDirection.rtl
+                  ? arabic.convert(span.text!)
+                  : useBidi && _textDirection == TextDirection.rtl
+                      ? bidi.logicalToVisual(span.text!)
+                      : span.text)!
               .split('\n');
 
           for (var line = 0; line < spanLines.length; line++) {
@@ -946,6 +955,32 @@ class RichText extends Widget with SpanningWidget {
 
               if (_softWrap &&
                   offsetX + metrics.width > constraintWidth + 0.00001) {
+                if (hyphenation != null) {
+                  final syllables = hyphenation!(word);
+                  if (syllables.length > 1) {
+                    var fits = '';
+                    for (var syllable in syllables) {
+                      if (offsetX +
+                              ((font.stringMetrics('$fits$syllable-',
+                                          letterSpacing: style.letterSpacing! /
+                                              (style.fontSize! *
+                                                  textScaleFactor)) *
+                                      (style.fontSize! * textScaleFactor))
+                                  .width) >
+                          constraintWidth + 0.00001) {
+                        break;
+                      }
+                      fits += syllable;
+                    }
+                    if (fits.isNotEmpty) {
+                      words[index] = '$fits-';
+                      words.insert(index + 1, word.substring(fits.length));
+                      index--;
+                      continue;
+                    }
+                  }
+                }
+
                 if (spanCount > 0 && metrics.width <= constraintWidth) {
                   overflow = true;
                   lines.add(_Line(
@@ -1041,7 +1076,8 @@ class RichText extends Widget with SpanningWidget {
               if (spanCount > 0) {
                 offsetY += bottom - top;
               } else {
-                offsetY += space.ascent + space.descent;
+                offsetY +=
+                    font.emptyLineHeight * style.fontSize! * textScaleFactor;
               }
               top = 0;
               bottom = 0;