+++ printing/windows/print_job.cpp
@@ -398,9 +398,8 @@ void PrintJob::rasterPdf(std::vector<uint8_t> data,
 
 std::map<std::string, bool> PrintJob::printingInfo() {
   return std::map<std::string, bool>{
-      {"directPrint", true},     {"dynamicLayout", true},   {"canPrint", true},
-      {"canListPrinters", true}, {"canConvertHtml", false}, {"canShare", true},
-      {"canRaster", true},
+      {"directPrint", true},     {"dynamicLayout", true}, {"canPrint", true},
+      {"canListPrinters", true}, {"canShare", true},      {"canRaster", true},
   };
 }