+++ pdf/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+# 1.2.0-napkin
+
+- Pull changes from DavBfr/dart_pdf
+
+## 3.11.0
+
+- Save in an isolate when available
+
+## 3.10.8
+
+- Add Flutter's Logical Pixel constant
+- Add support for existing reference objects
+- Update barcode golden pdf
+- Add support for hyphenation [ilja]
+- Add an option to disable bidirectional support [Olzhas-Suleimen]
+- Fix operator== type in TextDecoration class
+- Fixed wrong empty line height [janiselfert]
+- Add Support old Arabic method without bidi package [Baghdady92]
+
+## 3.10.7
+
+- Fix empty lines text gap
+- Fix lookup index of glyph for space character [Hendrik-Brower]
+
+## 3.10.6
+
+- Update bidi dependency
+
 # 1.1.0-napkin
 
 - Implement embedded svg images
@@ -19,6 +47,7 @@
 - Add RTL support [Milad-Akarie]
 - Fix Arabic fonts missing isolated form [Milad-Akarie]
 - Throw multi page error explicitly [Marcin Jeleński]
+- Fix deprecations
 
 ## 3.10.4